% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDID.R
\name{NDID}
\alias{NDID}
\title{NDID: Data Normalization and Differential Chromatin Interactions Detection from ChIA-PET Experiments}
\usage{
NDID(input_file, output_prefix)
}
\arguments{
\item{input_file, output_prefix}{}
}
\value{

}
\description{
NDID: Data Normalization and Differential Chromatin Interactions Detection from ChIA-PET Experiments
}
\examples{
NDID("input_file","output_prefix")
}
